rename worker manager package name to worker
